Sparse Brain
Implemented a basic version of a sparse neural network (not sure what to
call it, basically a neural net that isn't fully connected)

Currently not seeing any viable evolution.  One of the more difficult
things is that since it isn't fully connected there is no guarantee that
all the inputs/outputs will have a connection, namely birth and eating.

In this commit I've implemented "learning" which is just a "nongenetic"
set of features in the brain.  What I'm hoping is that in the early
stages of life it will help the creatures out, as well as give them a
limited shelf life and kill them off as they start living longer.
I'm going to test it on my desktop and hopefully there will be some
meaningful results when I awake.
If that doesn't happen I'll try a system where at least the inputs are
guaranteed an axon starting there and the outputs get an axon ending
there.

Also, before approving me to pull this back in someone remind me to
revert some of the rendering changes I made so I could see axons.